finish refactor PlanMateCli.kt
