+++ src/main/kotlin/ui/utils/ProjectSwimlaneUiPrinter.kt
@@ -1,4 +1,4 @@
-package net.thechance.ui.handlers
+package net.thechance.ui.utils
 
 import logic.entities.Project
 import ui.io.ConsoleIO
